+++ tools/fuzzer/CMakeLists.txt
@@ -12,24 +12,12 @@ if("${CMAKE_CURRENT_SOURCE_DIR}" STREQUAL "${CMAKE_CURRENT_BINARY_DIR}")
 	)
 endif()
 
+include(../../scripts/cmake/misc.cmake)
+
 include(ExternalProject)
 
 set(parser_src_dir "${PROJECT_SOURCE_DIR}/../tplg_parser")
 set(parser_install_dir "${PROJECT_BINARY_DIR}/sof_parser/install")
-
-ExternalProject_Add(sof_parser_ep
-	SOURCE_DIR "${parser_src_dir}"
-	PREFIX "${PROJECT_BINARY_DIR}/sof_parser"
-	BINARY_DIR "${PROJECT_BINARY_DIR}/sof_parser/build"
-	CMAKE_ARGS -DCMAKE_INSTALL_PREFIX=${parser_install_dir}
-		-DCMAKE_VERBOSE_MAKEFILE=${CMAKE_VERBOSE_MAKEFILE}
-	BUILD_ALWAYS 1
-	BUILD_BYPRODUCTS "${parser_install_dir}/lib/libsof_tplg_parser.so"
-)
-
-add_library(sof_parser SHARED IMPORTED)
-set_target_properties(sof_parser PROPERTIES IMPORTED_LOCATION "${parser_install_dir}/lib/libsof_tplg_parser.so")
-add_dependencies(sof_parser sof_parser_ep)
 
 add_executable(sof-fuzzer
 	main.c
@@ -40,8 +28,41 @@ add_executable(sof-fuzzer
 	platform/imx8-host.c
 )
 
+sof_append_relative_path_definitions(sof-fuzzer)
+
+target_compile_options(sof-fuzzer PRIVATE
+	-g -O -Wall -Werror -Wl,-EL -fPIC -DPIC
+	-Wmissing-prototypes -Wimplicit-fallthrough
+	-DCONFIG_LIBRARY -DCONFIG_IPC_MAJOR_3)
+
+set(SOF_ROOT_SOURCE_DIRECTORY "${PROJECT_SOURCE_DIR}/../..")
+
+target_include_directories(sof-fuzzer PRIVATE
+	"${SOF_ROOT_SOURCE_DIRECTORY}/src/include"
+	"${SOF_ROOT_SOURCE_DIRECTORY}/src/arch/host/include"
+	"${SOF_ROOT_SOURCE_DIRECTORY}/src/platform/library/include"
+	"${SOF_ROOT_SOURCE_DIRECTORY}"
+)
+
+set(config_h ${sof_binary_directory}/library_autoconfig.h)
+
 target_link_libraries(sof-fuzzer PRIVATE sof_parser)
 target_include_directories(sof-fuzzer PRIVATE "${parser_install_dir}/include")
+
+ExternalProject_Add(sof_parser_ep
+	SOURCE_DIR "${parser_src_dir}"
+	PREFIX "${PROJECT_BINARY_DIR}/sof_parser"
+	BINARY_DIR "${PROJECT_BINARY_DIR}/sof_parser/build"
+	CMAKE_ARGS
+		-DCMAKE_INSTALL_PREFIX=${PROJECT_BINARY_DIR}/sof_parser/install
+		-DCMAKE_VERBOSE_MAKEFILE=${CMAKE_VERBOSE_MAKEFILE}
+	BUILD_ALWAYS 1
+	BUILD_BYPRODUCTS "${parser_install_dir}/lib/libsof_tplg_parser.so"
+)
+
+add_library(sof_parser SHARED IMPORTED)
+set_target_properties(sof_parser PROPERTIES IMPORTED_LOCATION "${parser_install_dir}/lib/libsof_tplg_parser.so")
+add_dependencies(sof_parser sof_parser_ep)
 
 add_dependencies(sof-fuzzer sof_parser)
 
@@ -54,13 +75,5 @@ find_library(LIBRT rt)
 if(LIBRT)
 	target_link_libraries(sof-fuzzer PRIVATE ${LIBRT})
 endif()
-
-set(SOF_ROOT_SOURCE_DIRECTORY "${PROJECT_SOURCE_DIR}/../..")
-
-target_include_directories(sof-fuzzer PRIVATE
-	"${SOF_ROOT_SOURCE_DIRECTORY}/src/include"
-	"${SOF_ROOT_SOURCE_DIRECTORY}/src/arch/host/include"
-	"${SOF_ROOT_SOURCE_DIRECTORY}"
-)
 
 install(TARGETS sof-fuzzer DESTINATION bin)
