+++ tools/testbench/topology.c
@@ -18,8 +18,11 @@
 #include <sof/ipc/driver.h>
 #include <sof/ipc/topology.h>
 #include <tplg_parser/topology.h>
+#include <tplg_parser/tokens.h>
 #include "testbench/common_test.h"
 #include "testbench/file.h"
+
+#define MAX_TPLG_OBJECT_SIZE	4096
 
 const struct sof_dai_types sof_dais[] = {
 	{"SSP", SOF_DAI_INTEL_SSP},
@@ -94,6 +97,452 @@ void register_comp(int comp_type, struct sof_ipc_comp_ext *comp_ext)
 int find_widget(struct comp_info *temp_comp_list, int count, char *name)
 {
 	return 0;
+}
+
+/* load asrc dapm widget */
+int tplg_register_asrc(struct tplg_context *ctx)
+{
+	char tplg_object[MAX_TPLG_OBJECT_SIZE] = {0};
+	struct sof_ipc_comp *comp = (struct sof_ipc_comp *)tplg_object;
+	struct sof_ipc_comp_ext *comp_ext;
+	struct sof *sof = ctx->sof;
+	struct sof_ipc_comp_asrc *asrc;
+	int ret = 0;
+
+	ret = tplg_new_asrc(ctx, comp, MAX_TPLG_OBJECT_SIZE, NULL, 0);
+	if (ret < 0)
+		return ret;
+
+	asrc = (struct sof_ipc_comp_asrc *)comp;
+	comp_ext = (struct sof_ipc_comp_ext *)(asrc + 1);
+
+	/* set testbench input and output sample rate from topology */
+	if (!ctx->fs_out) {
+		ctx->fs_out = asrc->sink_rate;
+
+		if (!ctx->fs_in)
+			ctx->fs_in = asrc->source_rate;
+		else
+			asrc->source_rate = ctx->fs_in;
+	} else {
+		asrc->sink_rate = ctx->fs_out;
+	}
+
+	/* load asrc component */
+	register_comp(comp->type, comp_ext);
+	if (ipc_comp_new(sof->ipc, ipc_to_comp_new(&asrc)) < 0) {
+		fprintf(stderr, "error: new asrc comp\n");
+		return -EINVAL;
+	}
+
+	return ret;
+}
+
+/* load buffer DAPM widget */
+int tplg_register_buffer(struct tplg_context *ctx)
+{
+	struct sof *sof = ctx->sof;
+	struct sof_ipc_buffer buffer = {0};
+	int ret;
+
+	ret = tplg_create_buffer(ctx, &buffer);
+	if (ret < 0)
+		return ret;
+
+	if (tplg_create_controls(ctx->widget->num_kcontrols, ctx->file, NULL, 0) < 0) {
+		fprintf(stderr, "error: loading controls\n");
+		return -EINVAL;
+	}
+
+	/* create buffer component */
+	if (ipc_buffer_new(sof->ipc, &buffer) < 0) {
+		fprintf(stderr, "error: buffer new\n");
+		return -EINVAL;
+	}
+
+	return 0;
+}
+
+/* load pipeline graph DAPM widget*/
+int tplg_register_graph(void *dev, struct comp_info *temp_comp_list,
+			char *pipeline_string, FILE *file,
+			int count, int num_comps, int pipeline_id)
+{
+	struct sof_ipc_pipe_comp_connect connection;
+	struct sof *sof = (struct sof *)dev;
+	int ret = 0;
+	int i;
+
+	for (i = 0; i < count; i++) {
+		ret = tplg_create_graph(num_comps, pipeline_id, temp_comp_list,
+				      pipeline_string, &connection, file, i,
+				      count);
+		if (ret < 0)
+			return ret;
+
+		/* connect source and sink */
+		if (ipc_comp_connect(sof->ipc, ipc_to_pipe_connect(&connection)) < 0) {
+			fprintf(stderr, "error: comp connect\n");
+			return -EINVAL;
+		}
+	}
+
+	/* pipeline complete after pipeline connections are established */
+	for (i = 0; i < num_comps; i++) {
+		if (temp_comp_list[i].pipeline_id == pipeline_id &&
+		    temp_comp_list[i].type == SND_SOC_TPLG_DAPM_SCHEDULER)
+			ipc_pipeline_complete(sof->ipc, temp_comp_list[i].id);
+	}
+
+	return ret;
+}
+
+/* load mixer dapm widget */
+int tplg_register_mixer(struct tplg_context *ctx)
+{
+	char tplg_object[MAX_TPLG_OBJECT_SIZE] = {0};
+	struct sof_ipc_comp *comp = (struct sof_ipc_comp *)tplg_object;
+	struct sof_ipc_comp_ext *comp_ext;
+	struct sof_ipc_comp_mixer *mixer;
+	struct sof *sof = ctx->sof;
+	int ret = 0;
+
+	ret = tplg_new_mixer(ctx, comp, MAX_TPLG_OBJECT_SIZE, NULL, 0);
+	if (ret < 0)
+		return ret;
+
+	mixer = (struct sof_ipc_comp_mixer *)comp;
+	comp_ext = (struct sof_ipc_comp_ext *)(mixer + 1);
+
+	/* load mixer component */
+	register_comp(comp->type, comp_ext);
+	if (ipc_comp_new(sof->ipc, ipc_to_comp_new(comp)) < 0) {
+		fprintf(stderr, "error: new mixer comp\n");
+		ret = -EINVAL;
+	}
+
+	return ret;
+}
+
+int tplg_register_pga(struct tplg_context *ctx)
+{
+	char tplg_object[MAX_TPLG_OBJECT_SIZE] = {0};
+	struct sof_ipc_comp *comp = (struct sof_ipc_comp *)tplg_object;
+	struct sof_ipc_comp_ext *comp_ext;
+	struct sof_ipc_comp_volume *volume;
+	struct sof *sof = ctx->sof;
+	int ret;
+
+	ret = tplg_new_pga(ctx, comp, MAX_TPLG_OBJECT_SIZE, NULL, 0);
+	if (ret < 0) {
+		fprintf(stderr, "error: failed to create PGA\n");
+		return ret;
+	}
+
+	volume = (struct sof_ipc_comp_volume *)comp;
+	comp_ext = (struct sof_ipc_comp_ext *)(volume + 1);
+
+	/* load volume component */
+	register_comp(comp->type, comp_ext);
+	if (ipc_comp_new(sof->ipc, ipc_to_comp_new(comp)) < 0) {
+		fprintf(stderr, "error: new pga comp\n");
+		ret = -EINVAL;
+	}
+
+	return ret;
+}
+
+/* load scheduler dapm widget */
+int tplg_register_pipeline(struct tplg_context *ctx)
+{
+	struct sof *sof = ctx->sof;
+	struct sof_ipc_pipe_new pipeline = {0};
+	FILE *file = ctx->file;
+	int ret;
+
+	ret = tplg_create_pipeline(ctx, &pipeline);
+	if (ret < 0)
+		return ret;
+
+	if (tplg_create_controls(ctx->widget->num_kcontrols, file, NULL, 0) < 0) {
+		fprintf(stderr, "error: loading controls\n");
+		return -EINVAL;
+	}
+
+	pipeline.sched_id = ctx->sched_id;
+
+	/* Create pipeline */
+	if (ipc_pipeline_new(sof->ipc, (ipc_pipe_new *)&pipeline) < 0) {
+		fprintf(stderr, "error: pipeline new\n");
+		return -EINVAL;
+	}
+
+	return 0;
+}
+
+/* load process dapm widget */
+int load_process(struct tplg_context *ctx)
+{
+	struct sof *sof = ctx->sof;
+	struct snd_soc_tplg_dapm_widget *widget = ctx->widget;
+	struct sof_ipc_comp_process *process;
+	struct sof_ipc_comp_process *process_ipc = NULL;
+	struct snd_soc_tplg_ctl_hdr *ctl = NULL;
+	struct sof_ipc_comp_ext comp_ext;
+	char *priv_data = NULL;
+	int ret = 0;
+	int i;
+
+	process = malloc(sizeof(*process) + UUID_SIZE);
+	if (!process)
+		return -ENOMEM;
+
+	memset(process, 0, sizeof(*process) + UUID_SIZE);
+
+	ret = tplg_create_process(ctx, process, &comp_ext);
+	if (ret < 0)
+		return ret;
+
+	printf("debug uuid:\n");
+	for (i = 0; i < SOF_UUID_SIZE; i++)
+		printf("%u ", comp_ext.uuid[i]);
+	printf("\n");
+
+	/* Get control into ctl and priv_data */
+	for (i = 0; i < widget->num_kcontrols; i++) {
+		ret = tplg_create_single_control(&ctl, &priv_data, ctx->file);
+
+		if (ret < 0) {
+			fprintf(stderr, "error: failed control load\n");
+			return ret;
+		}
+
+		/* Merge process and priv_data into process_ipc */
+		if (priv_data)
+			ret = tplg_process_append_data(&process_ipc, process, ctl, priv_data);
+
+		free(ctl);
+		free(priv_data);
+		if (ret) {
+			fprintf(stderr, "error: private data append failed\n");
+			free(process_ipc);
+			return ret;
+		}
+	}
+
+	/* Default IPC without appended data */
+	if (!process_ipc) {
+		ret = tplg_process_init_data(&process_ipc, process);
+		if (ret)
+			return ret;
+	}
+
+	/* load process component */
+	register_comp(process_ipc->comp.type, &comp_ext);
+
+	/* Instantiate */
+	ret = ipc_comp_new(sof->ipc, ipc_to_comp_new(process_ipc));
+	free(process_ipc);
+	free(process);
+
+	if (ret < 0)
+		fprintf(stderr, "error: new process comp\n");
+
+	return ret;
+}
+
+/* load src dapm widget */
+int tplg_register_src(struct tplg_context *ctx)
+{
+	struct sof *sof = ctx->sof;
+	char tplg_object[MAX_TPLG_OBJECT_SIZE] = {0};
+	struct sof_ipc_comp *comp = (struct sof_ipc_comp *)tplg_object;
+	struct sof_ipc_comp_src *src;
+	struct sof_ipc_comp_ext *comp_ext;
+	int ret = 0;
+
+	ret = tplg_new_src(ctx, comp, MAX_TPLG_OBJECT_SIZE, NULL, 0);
+	if (ret < 0)
+		return ret;
+
+	src = (struct sof_ipc_comp_src *)comp;
+	comp_ext = (struct sof_ipc_comp_ext *)(src + 1);
+
+	/* set testbench input and output sample rate from topology */
+	if (!ctx->fs_out) {
+		ctx->fs_out = src->sink_rate;
+
+		if (!ctx->fs_in)
+			ctx->fs_in = src->source_rate;
+		else
+			src->source_rate = ctx->fs_in;
+	} else {
+		src->sink_rate = ctx->fs_out;
+	}
+
+	/* load src component */
+	register_comp(comp->type, comp_ext);
+	if (ipc_comp_new(sof->ipc, ipc_to_comp_new(comp)) < 0) {
+		fprintf(stderr, "error: new src comp\n");
+		return -EINVAL;
+	}
+
+	return ret;
+}
+
+/* load dapm widget */
+int load_widget(struct tplg_context *ctx)
+{
+	struct comp_info *temp_comp_list = ctx->info;
+	int comp_index = ctx->info_index;
+	int comp_id = ctx->comp_id;
+	int ret = 0;
+	int dev_type = ctx->dev_type;
+
+	if (!temp_comp_list) {
+		fprintf(stderr, "load_widget: temp_comp_list argument NULL\n");
+		return -EINVAL;
+	}
+
+	/* allocate memory for widget */
+	ctx->widget_size = sizeof(struct snd_soc_tplg_dapm_widget);
+	ctx->widget = malloc(ctx->widget_size);
+	if (!ctx->widget) {
+		fprintf(stderr, "error: mem alloc\n");
+		return -errno;
+	}
+
+	/* read widget data */
+	ret = fread(ctx->widget, ctx->widget_size, 1, ctx->file);
+	if (ret != 1) {
+		ret = -EINVAL;
+		goto exit;
+	}
+
+	/*
+	 * create a list with all widget info
+	 * containing mapping between component names and ids
+	 * which will be used for setting up component connections
+	 */
+	temp_comp_list[comp_index].id = comp_id;
+	temp_comp_list[comp_index].name = strdup(ctx->widget->name);
+	temp_comp_list[comp_index].type = ctx->widget->id;
+	temp_comp_list[comp_index].pipeline_id = ctx->pipeline_id;
+
+	printf("debug: loading comp_id %d: widget %s id %d\n",
+	       comp_id, ctx->widget->name, ctx->widget->id);
+
+	/* load widget based on type */
+	switch (ctx->widget->id) {
+
+	/* load pga widget */
+	case SND_SOC_TPLG_DAPM_PGA:
+		if (tplg_register_pga(ctx) < 0) {
+			fprintf(stderr, "error: load pga\n");
+			ret = -EINVAL;
+			goto exit;
+		}
+		break;
+	case SND_SOC_TPLG_DAPM_AIF_IN:
+		if (load_aif_in_out(ctx, SOF_IPC_STREAM_PLAYBACK) < 0) {
+			fprintf(stderr, "error: load AIF IN failed\n");
+			ret = -EINVAL;
+			goto exit;
+		}
+		break;
+	case SND_SOC_TPLG_DAPM_AIF_OUT:
+		if (load_aif_in_out(ctx, SOF_IPC_STREAM_CAPTURE) < 0) {
+			fprintf(stderr, "error: load AIF OUT failed\n");
+			ret = -EINVAL;
+			goto exit;
+		}
+		break;
+	case SND_SOC_TPLG_DAPM_DAI_IN:
+		if (load_dai_in_out(ctx, SOF_IPC_STREAM_PLAYBACK) < 0) {
+			fprintf(stderr, "error: load filewrite\n");
+			ret = -EINVAL;
+			goto exit;
+		}
+		break;
+	case SND_SOC_TPLG_DAPM_DAI_OUT:
+		if (load_dai_in_out(ctx, SOF_IPC_STREAM_CAPTURE) < 0) {
+			fprintf(stderr, "error: load filewrite\n");
+			ret = -EINVAL;
+			goto exit;
+		}
+		break;
+	case SND_SOC_TPLG_DAPM_BUFFER:
+		if (tplg_register_buffer(ctx) < 0) {
+			fprintf(stderr, "error: load buffer\n");
+			ret = -EINVAL;
+			goto exit;
+		}
+		break;
+
+	case SND_SOC_TPLG_DAPM_SCHEDULER:
+		/* find comp id for scheduling comp */
+		if (dev_type == FUZZER_DEV)
+			ctx->sched_id = find_widget(temp_comp_list, comp_id, ctx->widget->sname);
+
+		if (tplg_register_pipeline(ctx) < 0) {
+			fprintf(stderr, "error: load pipeline\n");
+			ret = -EINVAL;
+			goto exit;
+		}
+		break;
+
+	case SND_SOC_TPLG_DAPM_SRC:
+		if (tplg_register_src(ctx) < 0) {
+			fprintf(stderr, "error: load src\n");
+			ret = -EINVAL;
+			goto exit;
+		}
+		break;
+	case SND_SOC_TPLG_DAPM_ASRC:
+		if (tplg_register_asrc(ctx) < 0) {
+			fprintf(stderr, "error: load src\n");
+			ret = -EINVAL;
+			goto exit;
+		}
+		break;
+	case SND_SOC_TPLG_DAPM_MIXER:
+		if (tplg_register_mixer(ctx) < 0) {
+			fprintf(stderr, "error: load mixer\n");
+			ret = -EINVAL;
+			goto exit;
+		}
+		break;
+	case SND_SOC_TPLG_DAPM_EFFECT:
+		if (load_process(ctx) < 0) {
+			fprintf(stderr, "error: load effect\n");
+			ret = -EINVAL;
+			goto exit;
+		}
+		break;
+	/* unsupported widgets */
+	default:
+		if (fseek(ctx->file, ctx->widget->priv.size, SEEK_CUR)) {
+			fprintf(stderr, "error: fseek unsupported widget\n");
+			ret = -errno;
+			goto exit;
+		}
+
+		printf("info: Widget type not supported %d\n", ctx->widget->id);
+		ret = tplg_create_controls(ctx->widget->num_kcontrols, ctx->file, NULL, 0);
+		if (ret < 0) {
+			fprintf(stderr, "error: loading controls\n");
+			goto exit;
+		}
+		ret = 0;
+		break;
+	}
+
+	ret = 1;
+
+exit:
+	/* free allocated widget data */
+	free(ctx->widget);
+	return ret;
 }
 
 /* load fileread component */
@@ -240,7 +689,7 @@ static int load_fileread(struct tplg_context *ctx, int dir)
 	if (ret < 0)
 		return ret;
 
-	if (tplg_create_controls(ctx->widget->num_kcontrols, file) < 0) {
+	if (tplg_create_controls(ctx->widget->num_kcontrols, file, NULL, 0) < 0) {
 		fprintf(stderr, "error: loading controls\n");
 		return -EINVAL;
 	}
@@ -288,7 +737,7 @@ static int load_filewrite(struct tplg_context *ctx, int dir)
 	if (ret < 0)
 		return ret;
 
-	if (tplg_create_controls(ctx->widget->num_kcontrols, file) < 0) {
+	if (tplg_create_controls(ctx->widget->num_kcontrols, file, NULL, 0) < 0) {
 		fprintf(stderr, "error: loading controls\n");
 		return -EINVAL;
 	}
