+++ tools/tplg_parser/dai.c
@@ -13,14 +13,14 @@
 #include <errno.h>
 #include <unistd.h>
 #include <string.h>
-#include <math.h>
 #include <ipc/topology.h>
 #include <ipc/stream.h>
 #include <ipc/dai.h>
 #include <sof/common.h>
-#include <tplg_parser/topology.h>
 #include <sof/lib/uuid.h>
 #include <sof/ipc/topology.h>
+#include <tplg_parser/topology.h>
+#include <tplg_parser/tokens.h>
 
 static const struct sof_topology_token dai_tokens[] = {
 	{SOF_TKN_DAI_TYPE, SND_SOC_TPLG_TUPLE_TYPE_STRING, get_token_dai_type,
