+++ tools/tplg_parser/include/tplg_parser/tokens.h
@@ -1,0 +1,54 @@
+// SPDX-License-Identifier: BSD-3-Clause
+//
+// Copyright(c) 2019 Intel Corporation. All rights reserved.
+//
+// Author: Ranjani Sridharan <ranjani.sridharan@linux.intel.com>
+
+/* Topology parser */
+
+#ifndef _COMMON_TPLG_TOKENS_H
+#define _COMMON_TPLG_TOKENS_H
+
+#include <ipc/topology.h>
+#include <sof/ipc/topology.h>
+#include <tplg_parser/topology.h>
+
+/* temporary - current MAXLEN is not define in UAPI header - fix pending */
+#ifndef SNDRV_CTL_ELEM_ID_NAME_MAXLEN
+#define SNDRV_CTL_ELEM_ID_NAME_MAXLEN	44
+#endif
+
+#include <alsa/sound/uapi/asoc.h>
+
+struct sof_topology_token {
+	uint32_t token;
+	uint32_t type;
+	int (*get_token)(void *elem, void *object, uint32_t offset,
+			 uint32_t size);
+	uint32_t offset;
+	uint32_t size;
+};
+
+static const struct sof_topology_token tone_tokens[] = {
+};
+
+/* Generic components */
+static const struct sof_topology_token comp_tokens[] = {
+	{SOF_TKN_COMP_PERIOD_SINK_COUNT,
+		SND_SOC_TPLG_TUPLE_TYPE_WORD, get_token_uint32_t,
+		offsetof(struct sof_ipc_comp_config, periods_sink), 0},
+	{SOF_TKN_COMP_PERIOD_SOURCE_COUNT,
+		SND_SOC_TPLG_TUPLE_TYPE_WORD, get_token_uint32_t,
+		offsetof(struct sof_ipc_comp_config, periods_source), 0},
+	{SOF_TKN_COMP_FORMAT,
+		SND_SOC_TPLG_TUPLE_TYPE_STRING, get_token_comp_format,
+		offsetof(struct sof_ipc_comp_config, frame_fmt), 0},
+};
+
+/* Component extended tokens */
+static const struct sof_topology_token comp_ext_tokens[] = {
+	{SOF_TKN_COMP_UUID, SND_SOC_TPLG_TUPLE_TYPE_UUID, get_token_uuid, 0, 0},
+
+};
+
+#endif
