+++ tools/tplg_parser/pga.c
@@ -11,16 +11,13 @@
 #include <stdint.h>
 #include <stddef.h>
 #include <errno.h>
-#include <unistd.h>
 #include <string.h>
 #include <math.h>
 #include <ipc/topology.h>
-#include <ipc/stream.h>
-#include <ipc/dai.h>
-#include <sof/common.h>
 #include <sof/lib/uuid.h>
 #include <sof/ipc/topology.h>
 #include <tplg_parser/topology.h>
+#include <tplg_parser/tokens.h>
 
 /* volume */
 static const struct sof_topology_token volume_tokens[] = {
@@ -33,7 +30,8 @@ static const struct sof_topology_token volume_tokens[] = {
 };
 
 /* load pda dapm widget */
-int tplg_create_pga(struct tplg_context *ctx, struct sof_ipc_comp_volume *volume)
+int tplg_create_pga(struct tplg_context *ctx, struct sof_ipc_comp_volume *volume,
+		    size_t max_comp_size)
 {
 	struct snd_soc_tplg_vendor_array *array = NULL;
 	size_t total_array_size = 0, read_size;
@@ -42,6 +40,9 @@ int tplg_create_pga(struct tplg_context *ctx, struct sof_ipc_comp_volume *volume
 	int comp_id = ctx->comp_id;
 	char uuid[UUID_SIZE];
 	int ret;
+
+	if (max_comp_size < sizeof(struct sof_ipc_comp_volume) + UUID_SIZE)
+		return -EINVAL;
 
 	/* allocate memory for vendor tuple array */
 	array = (struct snd_soc_tplg_vendor_array *)malloc(size);
@@ -122,12 +123,11 @@ int tplg_create_pga(struct tplg_context *ctx, struct sof_ipc_comp_volume *volume
 }
 
 /* load pda dapm widget */
-int tplg_register_pga(struct tplg_context *ctx)
+int tplg_new_pga(struct tplg_context *ctx, struct sof_ipc_comp *comp, size_t comp_size,
+		struct snd_soc_tplg_ctl_hdr *rctl, size_t max_ctl_size)
 {
-	struct sof *sof = ctx->sof;
-	struct sof_ipc_comp *comp;
-	struct sof_ipc_comp_volume *volume;
 	struct snd_soc_tplg_ctl_hdr *ctl = NULL;
+	struct sof_ipc_comp_volume *volume = (struct sof_ipc_comp_volume *)comp;
 	struct snd_soc_tplg_mixer_control *mixer_ctl;
 	char *priv_data = NULL;
 	int32_t vol_min = 0;
@@ -138,15 +138,7 @@ int tplg_register_pga(struct tplg_context *ctx)
 	int channels = 0;
 	int ret = 0;
 
-	volume = malloc(sizeof(*volume) + UUID_SIZE);
-	if (!volume)
-		return -ENOMEM;
-
-	memset(volume, 0, sizeof(*volume) + UUID_SIZE);
-
-	comp = &volume->comp;
-
-	ret = tplg_create_pga(ctx, volume);
+	ret = tplg_create_pga(ctx, volume, comp_size);
 	if (ret < 0)
 		goto err;
 
@@ -171,6 +163,14 @@ int tplg_register_pga(struct tplg_context *ctx)
 		vol_step = mixer_ctl->hdr.tlv.scale.step;
 		vol_maxs = mixer_ctl->max;
 		channels = mixer_ctl->num_channels;
+
+		/* make sure the CTL will fit if we need to copy it for others */
+		if (max_ctl_size && ctl->size > max_ctl_size) {
+			fprintf(stderr, "error: failed pga control copy\n");
+			ret = -EINVAL;
+			goto out;
+		} else if (rctl)
+			memcpy(rctl, ctl, ctl->size);
 	}
 
 	vol_min_db = 0.01 * vol_min;
@@ -179,18 +179,10 @@ int tplg_register_pga(struct tplg_context *ctx)
 	volume->max_value = round(pow(10, vol_max_db / 20.0) * 65536);
 	volume->channels = channels;
 
+out:
 	free(ctl);
 	free(priv_data);
 
-	/* load volume component */
-	register_comp(comp->type, NULL);
-	if (ipc_comp_new(sof->ipc, ipc_to_comp_new(comp)) < 0) {
-		fprintf(stderr, "error: new pga comp\n");
-		ret = -EINVAL;
-		goto err;
-	}
-
 err:
-	free(volume);
 	return ret;
 }
